Add session_manager and config_manager
